refactor matches service with match generator modules
